1.2.9.7-alpha
changes:
- improved getter and setter in gameStates.
- now innateAbility has lvl 1, so you can upgrade to lvl 2.
- max abilities count = 5
- now abilities properly takes data from P_Stats

+++ Assets/Scripts/ActivationAbilities/RicochetStoneActivation.cs
@@ -10,7 +10,7 @@ public class RicochetStoneActivation : MonoBehaviour
     private Transform parentEnemy;
     private bool hasReduced = false; // Flag to track size reduction
 
-    public int currentDamage;
+    public float currentDamage;
 
     public void Start()
     {
